requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.6.0
pytz==2023.3
PyYAML==6.0
requests==2.30.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.2
